@@ -12,9 +12,10 @@
     "@nextui-org/react": "^2.2.10",
     "framer-motion": "^11.0.25",
     "lucide-react": "^0.365.0",
-    "next": "14.1.4",
+    "next": "^14.2.1",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-hot-toast": "^2.4.1"
   },
   "devDependencies": {
     "@types/node": "^20",
