@@ -1,10 +1,10 @@
 "use client"
 
-import { destroy } from "@/app/actions/categorias/destroy";
-import DropDown from "./DropDown";
-import { Icon } from "./Icon";
+import { Icon } from "@/components/Icon";
+import { destroy } from "../actions/categorias/destroy";
 import toast from "react-hot-toast";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
+import DropDown from "@/components/DropDown";
 
 interface CategoriaItemProps {
     categoria: Categoria
@@ -13,7 +13,7 @@ interface CategoriaItemProps {
 export function CategoriaItem(props: CategoriaItemProps) {
     const categoria = props.categoria
     const router = useRouter()
-    
+
     function handleDelete() {
         toast.promise(
             destroy(categoria.id),
@@ -24,16 +24,16 @@ export function CategoriaItem(props: CategoriaItemProps) {
             }
         );
     }
-    
+
     return (
         <div className="flex justify-between">
-            <div className="flex gap-2 border-white border-2 p-2 bg-black rounded-lg">
+            <div className="flex gap-2">
                 <Icon name={categoria.icone} />
                 <span>{categoria.nome}</span>
             </div>
             <DropDown 
                 onEdit={() => router.push("/categorias/" + categoria.id)}
-                onDelete={handleDelete}
+                onDelete={handleDelete} 
             />
         </div>
     )
