@@ -2,10 +2,8 @@
 
 import { redirect } from "next/navigation"
 
-export async function create(prevState: any, formData: FormData){
-
-    await new Promise(r => setTimeout(r, 1000))
-    
+export async function update(prevState: any, formData: FormData){
+    const id = formData.get("id")
 
     const data = {
         nome: formData.get("nome"),
@@ -15,7 +13,7 @@ export async function create(prevState: any, formData: FormData){
     }
 
     const options = {
-        method: "POST",
+        method: "PUT",
         body: JSON.stringify(data),
         headers: {
             "Content-Type": "application/json"
@@ -33,6 +31,4 @@ export async function create(prevState: any, formData: FormData){
             message: "campo nome é obrigatório"
         }
     }
-
-
 }
