@@ -5,7 +5,7 @@ import { SubmitButton } from "@/components/SubmitButton";
 import { Button, Input } from "@nextui-org/react";
 
 import { useFormState } from "react-dom";
-import { create } from "@/app/actions/categorias/create";
+import { update } from "@/app/actions/categorias/update";
 
 
 export function EditForm(categoria : Categoria) {
@@ -15,15 +15,17 @@ export function EditForm(categoria : Categoria) {
     }
 
 
-    const [state, formAction] = useFormState(create, initialState)
+    const [state, formAction] = useFormState(update, initialState)
     return (
         <form action={formAction} className="flex flex-col gap-4 p-6 m-6 bg-slate-900 rounded min-w-[500px]">
+            <input type="hidden" name="id" value={categoria.id} />
             <h2 className="text-2xl font-bold">Editar Categoria {categoria.nome}</h2>
             <Input
                 key="nome"
                 label="Nome"
                 name="nome"
                 variant="bordered"
+                defaultValue={categoria.nome}
                 labelPlacement="outside"
                 isInvalid={state?.message != ""}
                 errorMessage={state?.message}
@@ -34,6 +36,7 @@ export function EditForm(categoria : Categoria) {
                 label="Descrição"
                 name="descricao"
                 variant="bordered"
+                defaultValue={categoria.descricao}
                 labelPlacement="outside"
                 isInvalid={state?.message != ""}
                 errorMessage={state?.message}
@@ -44,6 +47,7 @@ export function EditForm(categoria : Categoria) {
                 label="Tipo"
                 name="tipo"
                 variant="bordered"
+                defaultValue={categoria.tipo}
                 labelPlacement="outside"
                 isInvalid={state?.message != ""}
                 errorMessage={state?.message}
@@ -54,6 +58,7 @@ export function EditForm(categoria : Categoria) {
                 label="Ícone"
                 name="icone"
                 variant="bordered"
+                defaultValue={categoria.icone}
                 labelPlacement="outside"
                 isInvalid={state?.message != ""}
                 errorMessage={state?.message}
