@@ -2,8 +2,14 @@
 
 import {Dropdown, DropdownTrigger, DropdownMenu, DropdownItem, Button} from "@nextui-org/react";
 import { ChevronDown, Pencil, X } from "lucide-react";
+import { MouseEventHandler } from "react";
 
-export default function DropDown() {
+interface DropDownActionsProps {
+  onEdit: MouseEventHandler,
+  onDelete: MouseEventHandler
+}
+
+export default function DropDownActions({onEdit, onDelete}: DropDownActionsProps) {
   return (
     <Dropdown>
       <DropdownTrigger>
@@ -14,8 +20,8 @@ export default function DropDown() {
         </Button>
       </DropdownTrigger>
       <DropdownMenu aria-label="Static Actions">
-        <DropdownItem key="edit" startContent={<Pencil size={18}/>}>Editar</DropdownItem>
-        <DropdownItem key="delete" className="text-danger" color="danger" startContent={<X size={20}/>}>
+        <DropdownItem onClick={onEdit} key="edit" startContent={<Pencil size={18}/>}>Editar</DropdownItem>
+        <DropdownItem onClick={onDelete} key="delete" className="text-danger" color="danger" startContent={<X size={20}/>}>
           Apagar
         </DropdownItem>
       </DropdownMenu>
